move solveLS to torchstain.tf.utils
