+++ torchstain/tf/utils/__init__.py
@@ -1,3 +1,3 @@
 from torchstain.tf.utils.cov import cov
 from torchstain.tf.utils.percentile import percentile
-from torchstain.tf.normalizers.solveLS import solveLS
+from torchstain.tf.utils.solveLS import solveLS
